+++ app/index.html
@@ -27,10 +27,7 @@
     <!-- bluebird -->
     <script src="vendor/js/bluebird.min.js" defer></script>
     <!-- Custom JS -->
-    <script src="js/layout.js" defer></script>
-    <script src="js/error.js" defer></script>
-    <script src="js/ethnet.js" defer></script>
-    <script src="js/app.js" defer></script>
+    <script src="js/bundle.js" defer></script>
   </head>
   <body>
     <!-- Content -->
