+++ scripts/build.sh
@@ -9,3 +9,6 @@ rm -f app/contracts/*
 
 # Copy the newly compiled artifacts to the app's directory
 rsync build/contracts/* app/contracts/ --exclude=Migrations.json
+
+# Bundle the app's JavaScript sources
+browserify app/js/src/app.js -o app/js/bundle.js
